Change the code in the tree

node_replace() stores the replacement for a particular node, it doesn't change the actual code. node_replace_all() does the same but for several nodes (e.g. the output of node_find_all()).

node_commit_edits() takes a list of replacements (the output of node_replace() or node_replace_all()) and applies them one by one to the input node, returning the modified code.

Usage

node_replace(x, new_text)

node_replace_all(x, new_text)

node_commit_edits(x, edits)

Arguments

x

A node, either from tree_root() or from another node_*() function.

new_text

The replacement for a node.

edits

A list of replacements (the output of node_replace() or node_replace_all()).

Examples

src <- "x <- rnorm(100, mean = 2)
any(duplicated(y))
plot(mtcars)
any(duplicated(x))"

root <- src |>
  tree_new() |>
  tree_root()

# one replacement ------------------------------------------

node_to_fix <- root |>
  node_find(pattern = "any(duplicated($A))")

fix <- node_to_fix |>
  node_replace(
    paste0(
      "anyDuplicated(",
      node_text(node_get_match(node_to_fix, "A")),
      ") > 0"
    )
  )

node_commit_edits(root, fix) |>
  cat()
#> x <- rnorm(100, mean = 2)
#> anyDuplicated(y) > 0
#> plot(mtcars)
#> any(duplicated(x))

# several replacements ------------------------------------------

nodes_to_fix <- root |>
  node_find_all(pattern = "any(duplicated($A))")

fixes <- nodes_to_fix |>
  node_replace_all(
    paste0(
      "anyDuplicated(",
      node_text_all(lapply(nodes_to_fix, function(x) node_get_match(x, "A"))),
      ") > 0"
    )
  )

node_commit_edits(root, fixes) |>
  cat()
#> x <- rnorm(100, mean = 2)
#> anyDuplicated(y) > 0
#> plot(mtcars)
#> anyDuplicated(x) > 0